+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-05-14 13:59:20
-// Commit: feat(core): add bi-directional event bus handler for core
+// Updated at: 2026-05-14 18:30:06
+// Commit: feat(core): add async core stream provider
 
 class coreComponentState {
-  final String commitRef = '1140';
-  final String lastMessage = 'feat(core): add bi-directional event bus handler for core';
-  final DateTime timestamp = DateTime.parse('2026-05-14T13:59:20+05:00');
+  final String commitRef = '1156';
+  final String lastMessage = 'feat(core): add async core stream provider';
+  final DateTime timestamp = DateTime.parse('2026-05-14T18:30:06+05:00');
 }